Move project relevant objects to own package
